Add mount_point demo to the showcase
