+++ examples/showcase/sub/mount_point/Cargo.toml
@@ -5,4 +5,4 @@ authors = ["Ben Berman <ben@standardbots.com>"]
 
 [dependencies]
 stdweb = "0.4"
-yew = { path = "../.." }
+yew = { path = "../../../.." }
